+++ packages/jaspr_builder/lib/src/codec/codecs.dart
@@ -84,7 +84,7 @@ class DecoderVisitor extends UnifyingTypeVisitorWithArgument<String?, String> {
   @override
   String? visitDartType(DartType type, String argument) {
     if (!type.isDartPrimitive) {
-      throw UnsupportedError('Unsupported parameter type: Expected primitive type, found ${type.getDisplayString()}');
+      throw const InvalidParameterException();
     }
     return argument;
   }
@@ -105,8 +105,7 @@ class DecoderVisitor extends UnifyingTypeVisitorWithArgument<String?, String> {
       }
     } else if (type.isDartCoreMap) {
       if (!type.typeArguments.first.isDartCoreString) {
-        throw UnsupportedError(
-            'Unsupported Map key type: Expected String, found ${type.typeArguments.first.getDisplayString()}');
+        throw const InvalidParameterException();
       }
 
       var nullCheck = type.nullabilitySuffix == NullabilitySuffix.question ? '?' : '';
@@ -166,4 +165,8 @@ extension on DartType {
             (this as InterfaceType).typeArguments.first.isDartCoreString &&
             (this as InterfaceType).typeArguments.last is DynamicType);
   }
+}
+
+class InvalidParameterException implements Exception {
+  const InvalidParameterException();
 }