+++ packages/jaspr_builder/lib/src/sync/sync_mixins_builder.dart
@@ -2,7 +2,7 @@
 
 import 'dart:async';
 
-import 'package:analyzer/dart/element/element.dart';
+import 'package:analyzer/dart/element/element2.dart';
 import 'package:build/build.dart';
 
 import '../codec/codecs.dart';
@@ -46,24 +46,23 @@ class SyncMixinsBuilder implements Builder {
 
     var library = await buildStep.inputLibrary;
 
-    var annotated = library.topLevelElements
-        .whereType<ClassElement>()
+    var annotated = library.classes
         .map((clazz) => (
               clazz,
-              clazz.fields.where((element) => syncChecker.firstAnnotationOfExact(element) != null).where((element) {
+              clazz.fields2.where((element) => syncChecker.firstAnnotationOfExact(element) != null).where((element) {
                 if (element.isStatic) {
                   log.severe(
-                      '@sync cannot be used on static fields. Failing element: ${clazz.name}.${element.name} in library ${library.source.fullName}.');
+                      '@sync cannot be used on static fields. Failing element: ${clazz.name3}.${element.name3} in library ${library.firstFragment.source.fullName}.');
                   return false;
                 }
                 if (element.isFinal) {
                   log.severe(
-                      '@sync cannot be used on final fields. Failing element: ${clazz.name}.${element.name} in library ${library.source.fullName}.');
+                      '@sync cannot be used on final fields. Failing element: ${clazz.name3}.${element.name3} in library ${library.firstFragment.source.fullName}.');
                   return false;
                 }
                 if (element.isPrivate) {
                   log.severe(
-                      '@sync cannot be used on private fields. Failing element: ${clazz.name}.${element.name} in library ${library.source.fullName}.');
+                      '@sync cannot be used on private fields. Failing element: ${clazz.name3}.${element.name3} in library ${library.firstFragment.source.fullName}.');
                   return false;
                 }
 
@@ -74,7 +73,7 @@ class SyncMixinsBuilder implements Builder {
         .where((c) {
       if (!stateChecker.isSuperOf(c.$1)) {
         log.severe(
-            '@sync can only be used on fields in a State class. Failing element: ${c.$1.name} in library ${library.source.fullName}.');
+            '@sync can only be used on fields in a State class. Failing element: ${c.$1.name3} in library ${library.firstFragment.source.fullName}.');
         return false;
       }
       return true;
@@ -105,30 +104,36 @@ class SyncMixinsBuilder implements Builder {
     await buildStep.writeAsFormattedDart(outputId, source);
   }
 
-  String generateMixinFromEntry((ClassElement, Iterable<FieldElement>) element, Codecs codecs, String baseImport) {
+  String generateMixinFromEntry((ClassElement2, Iterable<FieldElement2>) element, Codecs codecs, String baseImport) {
     final (clazz, fields) = element;
     final comp = clazz.supertype!.typeArguments.first.element!;
 
     final members = fields.map((f) {
       var type = codecs.getPrefixedType(f.type);
       return """
-        $type get ${f.name};
-        set ${f.name}($type ${f.name});
+        $type get ${f.name3};
+        set ${f.name3}($type ${f.name3});
       """;
     }).join('\n');
 
     final decoders = fields.map((f) {
-      var decoder = codecs.getDecoderFor(f.type, "value['${f.name}']");
-      return "${f.name} = $decoder;";
+      try {
+        var decoder = codecs.getDecoderFor(f.type, "value['${f.name3}']");
+        return "${f.name3} = $decoder;";
+      } on InvalidParameterException catch (_) {
+        throw UnsupportedError(
+            'Fields annotated with @sync must have a primitive serializable type or a type that defines @decoder and @encoder methods. '
+            'Failing field: [$f] in ${clazz.name3}');
+      }
     }).join('\n');
 
     final encoders = fields.map((f) {
-      var encoder = codecs.getEncoderFor(f.type, f.name);
-      return "'${f.name}': $encoder,";
+      var encoder = codecs.getEncoderFor(f.type, f.name3 ?? '');
+      return "'${f.name3}': $encoder,";
     }).join('\n');
 
     return """
-      mixin ${clazz.name.startsWith('_') ? clazz.name.substring(1) : clazz.name}SyncMixin on State<[[$baseImport]].${comp.name}> implements SyncStateMixin<[[$baseImport]].${comp.name}, Map<String, dynamic>> {
+      mixin ${clazz.name3?.startsWith('_') ?? false ? clazz.name3!.substring(1) : clazz.name3}SyncMixin on State<[[$baseImport]].${comp.name}> implements SyncStateMixin<[[$baseImport]].${comp.name}, Map<String, dynamic>> {
         $members
       
         @override