+++ packages/jaspr_builder/test/client/client_annotation_test.dart
@@ -12,6 +12,13 @@ import 'sources/client_model_extension.dart';
 
 void main() {
   group('client annotation', () {
+    late TestReaderWriter reader;
+
+    setUp(() async {
+      reader = TestReaderWriter(rootPackage: 'models');
+      await reader.testing.loadIsolateSources();
+    });
+
     group('on basic component', () {
       test('generates json module', () async {
         await testBuilder(
@@ -21,7 +28,7 @@ void main() {
             ...clientBasicJsonOutputs,
             'site|lib/component_basic.client.dart': isNotEmpty,
           },
-          reader: await PackageAssetReader.currentIsolate(),
+          readerWriter: reader,
         );
       });
 
@@ -33,7 +40,7 @@ void main() {
             'site|lib/component_basic.client.json': isNotEmpty,
             ...clientBasicDartOutputs,
           },
-          reader: await PackageAssetReader.currentIsolate(),
+          readerWriter: reader,
         );
       });
     });
@@ -47,7 +54,7 @@ void main() {
             ...clientModelClassJsonOutputs,
             'site|lib/component_model_class.client.dart': isNotEmpty,
           },
-          reader: await PackageAssetReader.currentIsolate(),
+          readerWriter: reader,
         );
       });
 
@@ -59,7 +66,7 @@ void main() {
             'site|lib/component_model_class.client.json': isNotEmpty,
             ...clientModelClassDartOutputs,
           },
-          reader: await PackageAssetReader.currentIsolate(),
+          readerWriter: reader,
         );
       });
     });
@@ -73,7 +80,7 @@ void main() {
             ...clientModelExtensionJsonOutputs,
             'site|lib/component_model_extension.client.dart': isNotEmpty,
           },
-          reader: await PackageAssetReader.currentIsolate(),
+          readerWriter: reader,
         );
       });
 
@@ -85,7 +92,7 @@ void main() {
             'site|lib/component_model_extension.client.json': isNotEmpty,
             ...clientModelExtensionDartOutputs,
           },
-          reader: await PackageAssetReader.currentIsolate(),
+          readerWriter: reader,
         );
       });
     });
@@ -98,18 +105,18 @@ void main() {
           ClientModuleBuilder(BuilderOptions({})),
           clientInvalidConstructorSources,
           outputs: {},
-          reader: await PackageAssetReader.currentIsolate(),
           onLog: (log) {
             if (log.level.name == 'SEVERE') {
               errorLog = log.message;
             }
           },
+          readerWriter: reader,
         );
 
         expect(
           errorLog,
-          equals('Client components only support initializing formal constructor parameters. '
-              'Failing element: Component.(String a)'),
+          equals('ClientModuleBuilder on lib/component_basic.dart:\nClient components only support initializing formal constructor parameters. '
+              'Failing element: Component.new(String a)'),
         );
       });
 
@@ -120,17 +127,17 @@ void main() {
           ClientModuleBuilder(BuilderOptions({})),
           clientInvalidParamSources,
           outputs: {},
-          reader: await PackageAssetReader.currentIsolate(),
           onLog: (log) {
             if (log.level.name == 'SEVERE') {
               errorLog = log.message;
             }
           },
+          readerWriter: reader,
         );
 
         expect(
           errorLog,
-          equals('Unsupported parameter type: Expected primitive type, found DateTime'),
+          equals('ClientModuleBuilder on lib/component_basic.dart:\n@client components only support parameters of primitive serializable types or types that define @decoder and @encoder methods. Failing parameter: [DateTime time] in Component.new()'),
         );
       });
     });
@@ -144,7 +151,7 @@ void main() {
           ...clientModelExtensionJsonOutputs,
         },
         outputs: clientBundleOutputs,
-        reader: await PackageAssetReader.currentIsolate(),
+        readerWriter: reader,
       );
     });
   });